+++ launchDaemonManagement.zsh
@@ -21,15 +21,16 @@
 #
 # HISTORY
 #
-# Version 2.1.0b10, 12-Dec-2025, Dan K. Snelson (@dan-snelson)
+# Version 2.1.0b11, 12-Dec-2025, Dan K. Snelson (@dan-snelson)
 #   - Added ability to use `titleMessageUpdateOrUpgrade:l` (Pull Request #26; thanks, @maxsundellacne!)
 #   - Added logic to hide `button2` based on `DaysBeforeDeadlineHidingButton2` (Pull Request #27; thanks, @maxsundellacne!)
 #   - Refactored `resetConfiguration` function to avoid errors when attempting to `chmod` non-existent files
-#   - Added warning for excessive uptime (configurable via `DaysOfExcessiveUptimeWarning` variable; #28)
+#   - Added warning for excessive uptime (configurable via `DaysOfExcessiveUptimeWarning` variable; Feature Request #28)
 #   - Added logic for when the reminder dialog is re-displayed after clicking the `infobutton` (based on if we're already hiding the secondary button; #31)
 #   - Streamline Deployment & Documentation (Feature Request #35)
 #   - Addressed Bugs #34 (thanks, @TechTrekkie!) and #36 (I. Blame. AI.)
 #   - Refactored `assemble.zsh` (thanks for the feedback, @Andrew!)
+#   - Added warning for low disk space (configurable via `minimumDiskFreePercentage` variable; Feature Request #39. (Thanks for the suggestion, @prgsenright!)
 #
 ####################################################################################################
 
@@ -44,7 +45,7 @@
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local:/usr/local/bin
 
 # Script Version
-scriptVersion="2.1.0b10"
+scriptVersion="2.1.0b11"
 
 # Client-side Log
 scriptLog="/var/log/org.churchofjesuschrist.log"