+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Changelog
 
-### Version 2.1.0b10 (12-Dec-2025)
+### Version 2.1.0b11 (12-Dec-2025)
 - Added ability to use `titleMessageUpdateOrUpgrade:l` (Pull Request #26; thanks, @maxsundellacne!)
 - Added logic to hide `button2` based on `DaysBeforeDeadlineHidingButton2` (Pull Request #27; thanks, @maxsundellacne!)
 - Refactored `resetConfiguration` function to avoid errors when attempting to `chmod` non-existent files
@@ -12,6 +12,8 @@
 - Streamline Deployment & Documentation (Feature Request #35)
 - Addressed Bugs #34 (thanks, @TechTrekkie!) and #36 (I. Blame. AI.)
 - Refactored `assemble.zsh` (thanks for the feedback, @Andrew!)
+- Added warning for low disk space (configurable via `minimumDiskFreePercentage` variable; Feature Request #39. (Thanks for the suggestion, @prgsenright!)
+
 
 ### Version 2.0.0 (06-Dec-2025)
 - Reorganized script structure for (hopefully) improved clarity